Split Collections controller
