+++ src/controllers/collections/Bookmarks.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace flusio\controllers\collections;
+
+use Minz\Response;
+use flusio\models;
+use flusio\utils;
+
+/**
+ * @author  Marien Fressinaud <dev@marienfressinaud.fr>
+ * @license http://www.gnu.org/licenses/agpl-3.0.en.html AGPL
+ */
+class Bookmarks
+{
+    /**
+     * Show the bookmarks page
+     *
+     * @response 302 /login?redirect_to=/bookmarks if not connected
+     * @response 404 if the collection doesn’t exist or user hasn't access
+     * @response 200
+     */
+    public function show()
+    {
+        $user = utils\CurrentUser::get();
+        if (!$user) {
+            return Response::redirect('login', [
+                'redirect_to' => \Minz\Url::for('bookmarks'),
+            ]);
+        }
+
+        $bookmarks = $user->bookmarks();
+        if (!$bookmarks) {
+            \Minz\Log::error("User {$user->id} has no Bookmarks collection.");
+            return Response::notFound('not_found.phtml', [
+                'details' => _('It looks like you have no “Bookmarks” collection, you should contact the support.'),
+            ]);
+        }
+
+        return Response::ok('collections/bookmarks/show.phtml', [
+            'collection' => $bookmarks,
+            'links' => $bookmarks->links(),
+        ]);
+    }
+}
