+++ src/views/collections/discovery/show.phtml
