+++ tests/controllers/CollectionsTest.php
@@ -861,466 +861,4 @@ class CollectionsTest extends \PHPUnit\Framework\TestCase
         $this->assertTrue(models\Collection::exists($collection_id));
         $this->assertFlash('error', 'A security verification failed.');
     }
-
-    public function testShowBookmarksRendersCorrectly()
-    {
-        $user = $this->login();
-        $link_title = $this->fake('words', 3, true);
-        $collection_id = $this->create('collection', [
-            'user_id' => $user->id,
-            'type' => 'bookmarks',
-        ]);
-        $link_id = $this->create('link', [
-            'user_id' => $user->id,
-            'title' => $link_title,
-        ]);
-        $this->create('link_to_collection', [
-            'link_id' => $link_id,
-            'collection_id' => $collection_id,
-        ]);
-
-        $response = $this->appRun('get', '/bookmarks');
-
-        $this->assertResponse($response, 200, $link_title);
-        $this->assertPointer($response, 'collections/show_bookmarks.phtml');
-    }
-
-    public function testShowBookmarksRedirectsIfNotConnected()
-    {
-        $response = $this->appRun('get', '/bookmarks');
-
-        $this->assertResponse($response, 302, '/login?redirect_to=%2Fbookmarks');
-    }
-
-    public function testShowBookmarksFailsIfCollectionDoesNotExist()
-    {
-        $this->login();
-
-        $response = $this->appRun('get', '/bookmarks');
-
-        $this->assertResponse($response, 404, 'It looks like you have no “Bookmarks” collection');
-    }
-
-    public function testDiscoverRendersCorrectly()
-    {
-        $user = $this->login();
-        $collection_name = $this->fake('sentence');
-        $other_user_id = $this->create('user');
-        $collection_id = $this->create('collection', [
-            'type' => 'collection',
-            'user_id' => $other_user_id,
-            'name' => $collection_name,
-            'is_public' => 1,
-        ]);
-        $link_id = $this->create('link', [
-            'user_id' => $other_user_id,
-            'is_hidden' => 0,
-        ]);
-        $this->create('link_to_collection', [
-            'collection_id' => $collection_id,
-            'link_id' => $link_id,
-        ]);
-
-        $response = $this->appRun('get', '/collections/discover');
-
-        $this->assertResponse($response, 200, $collection_name);
-        $this->assertPointer($response, 'collections/discover.phtml');
-    }
-
-    public function testDiscoverDoesNotListOwnedCollections()
-    {
-        $user = $this->login();
-        $collection_name = $this->fake('sentence');
-        $collection_id = $this->create('collection', [
-            'type' => 'collection',
-            'user_id' => $user->id,
-            'name' => $collection_name,
-            'is_public' => 1,
-        ]);
-        $link_id = $this->create('link', [
-            'user_id' => $user->id,
-            'is_hidden' => 0,
-        ]);
-        $this->create('link_to_collection', [
-            'collection_id' => $collection_id,
-            'link_id' => $link_id,
-        ]);
-
-        $response = $this->appRun('get', '/collections/discover');
-
-        $output = $response->render();
-        $this->assertStringNotContainsString($collection_name, $output);
-    }
-
-    public function testDiscoverDoesNotListEmptyCollections()
-    {
-        $user = $this->login();
-        $collection_name = $this->fake('sentence');
-        $other_user_id = $this->create('user');
-        $collection_id = $this->create('collection', [
-            'type' => 'collection',
-            'user_id' => $other_user_id,
-            'name' => $collection_name,
-            'is_public' => 1,
-        ]);
-
-        $response = $this->appRun('get', '/collections/discover');
-
-        $output = $response->render();
-        $this->assertStringNotContainsString($collection_name, $output);
-    }
-
-    public function testDiscoverDoesNotListPrivateCollections()
-    {
-        $user = $this->login();
-        $collection_name = $this->fake('sentence');
-        $other_user_id = $this->create('user');
-        $collection_id = $this->create('collection', [
-            'type' => 'collection',
-            'user_id' => $other_user_id,
-            'name' => $collection_name,
-            'is_public' => 0,
-        ]);
-        $link_id = $this->create('link', [
-            'user_id' => $other_user_id,
-            'is_hidden' => 0,
-        ]);
-        $this->create('link_to_collection', [
-            'collection_id' => $collection_id,
-            'link_id' => $link_id,
-        ]);
-
-        $response = $this->appRun('get', '/collections/discover');
-
-        $output = $response->render();
-        $this->assertStringNotContainsString($collection_name, $output);
-    }
-
-    public function testDiscoverDoesNotListFeedCollections()
-    {
-        $user = $this->login();
-        $collection_name = $this->fake('sentence');
-        $other_user_id = $this->create('user');
-        $collection_id = $this->create('collection', [
-            'type' => 'feed',
-            'user_id' => $other_user_id,
-            'name' => $collection_name,
-            'is_public' => 1,
-        ]);
-        $link_id = $this->create('link', [
-            'user_id' => $other_user_id,
-            'is_hidden' => 0,
-        ]);
-        $this->create('link_to_collection', [
-            'collection_id' => $collection_id,
-            'link_id' => $link_id,
-        ]);
-
-        $response = $this->appRun('get', '/collections/discover');
-
-        $output = $response->render();
-        $this->assertStringNotContainsString($collection_name, $output);
-    }
-
-    public function testDiscoverDoesNotCountHiddenLinks()
-    {
-        $user = $this->login();
-        $collection_name = $this->fake('sentence');
-        $other_user_id = $this->create('user');
-        $collection_id = $this->create('collection', [
-            'type' => 'collection',
-            'user_id' => $other_user_id,
-            'name' => $collection_name,
-            'is_public' => 1,
-        ]);
-        $link_id1 = $this->create('link', [
-            'user_id' => $other_user_id,
-            'is_hidden' => 0,
-        ]);
-        $link_id2 = $this->create('link', [
-            'user_id' => $other_user_id,
-            'is_hidden' => 1,
-        ]);
-        $this->create('link_to_collection', [
-            'collection_id' => $collection_id,
-            'link_id' => $link_id1,
-        ]);
-        $this->create('link_to_collection', [
-            'collection_id' => $collection_id,
-            'link_id' => $link_id2,
-        ]);
-
-        $response = $this->appRun('get', '/collections/discover');
-
-        $output = $response->render();
-        $this->assertStringContainsString('1 link', $output);
-        $this->assertStringNotContainsString('2 links', $output);
-    }
-
-    public function testDiscoverRedirectsIfPageIsOutOfBound()
-    {
-        $user = $this->login();
-        $collection_name = $this->fake('sentence');
-        $other_user_id = $this->create('user');
-        $collection_id = $this->create('collection', [
-            'type' => 'collection',
-            'user_id' => $other_user_id,
-            'name' => $collection_name,
-            'is_public' => 1,
-        ]);
-        $link_id = $this->create('link', [
-            'user_id' => $other_user_id,
-            'is_hidden' => 0,
-        ]);
-        $this->create('link_to_collection', [
-            'collection_id' => $collection_id,
-            'link_id' => $link_id,
-        ]);
-
-        $response = $this->appRun('get', '/collections/discover', [
-            'page' => 0,
-        ]);
-
-        $this->assertResponse($response, 302, '/collections/discover?page=1');
-    }
-
-    public function testDiscoverRedirectsIfNotConnected()
-    {
-        $collection_name = $this->fake('sentence');
-        $other_user_id = $this->create('user');
-        $collection_id = $this->create('collection', [
-            'type' => 'collection',
-            'user_id' => $other_user_id,
-            'name' => $collection_name,
-            'is_public' => 1,
-        ]);
-        $link_id = $this->create('link', [
-            'user_id' => $other_user_id,
-            'is_hidden' => 0,
-        ]);
-        $this->create('link_to_collection', [
-            'collection_id' => $collection_id,
-            'link_id' => $link_id,
-        ]);
-
-        $response = $this->appRun('get', '/collections/discover');
-
-        $this->assertResponse($response, 302, "/login?redirect_to=%2Fcollections%2Fdiscover");
-    }
-
-    public function testFollowMakesUserFollowingAndRedirects()
-    {
-        $user = $this->login();
-        $owner_id = $this->create('user');
-        $followed_collection_dao = new models\dao\FollowedCollection();
-        $collection_id = $this->create('collection', [
-            'user_id' => $owner_id,
-            'type' => 'collection',
-            'is_public' => 1,
-        ]);
-
-        $this->assertSame(0, $followed_collection_dao->count());
-
-        $response = $this->appRun('post', "/collections/{$collection_id}/follow", [
-            'csrf' => $user->csrf,
-        ]);
-
-        $this->assertResponse($response, 302, "/collections/{$collection_id}");
-        $this->assertSame(1, $followed_collection_dao->count());
-        $db_followed_collection = $followed_collection_dao->listAll()[0];
-        $this->assertSame($user->id, $db_followed_collection['user_id']);
-        $this->assertSame($collection_id, $db_followed_collection['collection_id']);
-    }
-
-    public function testFollowRedirectsIfNotConnected()
-    {
-        $owner_id = $this->create('user');
-        $followed_collection_dao = new models\dao\FollowedCollection();
-        $collection_id = $this->create('collection', [
-            'user_id' => $owner_id,
-            'type' => 'collection',
-            'is_public' => 1,
-        ]);
-
-        $response = $this->appRun('post', "/collections/{$collection_id}/follow", [
-            'csrf' => 'a token',
-        ]);
-
-        $this->assertResponse($response, 302, "/login?redirect_to=%2Fcollections%2F{$collection_id}");
-        $this->assertSame(0, $followed_collection_dao->count());
-    }
-
-    public function testFollowFailsIfCollectionDoesNotExist()
-    {
-        $user = $this->login();
-        $owner_id = $this->create('user');
-        $followed_collection_dao = new models\dao\FollowedCollection();
-        $collection_id = $this->create('collection', [
-            'user_id' => $owner_id,
-            'type' => 'collection',
-            'is_public' => 1,
-        ]);
-
-        $response = $this->appRun('post', '/collections/unknown/follow', [
-            'csrf' => $user->csrf,
-        ]);
-
-        $this->assertResponse($response, 404);
-        $this->assertSame(0, $followed_collection_dao->count());
-    }
-
-    public function testFollowFailsIfUserHasNoAccess()
-    {
-        $user = $this->login();
-        $owner_id = $this->create('user');
-        $followed_collection_dao = new models\dao\FollowedCollection();
-        $collection_id = $this->create('collection', [
-            'user_id' => $owner_id,
-            'type' => 'collection',
-            'is_public' => 0,
-        ]);
-
-        $response = $this->appRun('post', "/collections/{$collection_id}/follow", [
-            'csrf' => $user->csrf,
-        ]);
-
-        $this->assertResponse($response, 404);
-        $this->assertSame(0, $followed_collection_dao->count());
-    }
-
-    public function testFollowFailsIfCsrfIsInvalid()
-    {
-        $user = $this->login();
-        $owner_id = $this->create('user');
-        $followed_collection_dao = new models\dao\FollowedCollection();
-        $collection_id = $this->create('collection', [
-            'user_id' => $owner_id,
-            'type' => 'collection',
-            'is_public' => 1,
-        ]);
-
-        $response = $this->appRun('post', "/collections/{$collection_id}/follow", [
-            'csrf' => 'not the token',
-        ]);
-
-        $this->assertResponse($response, 302, "/collections/{$collection_id}");
-        $this->assertFlash('error', 'A security verification failed: you should retry to submit the form.');
-        $this->assertSame(0, $followed_collection_dao->count());
-    }
-
-    public function testUnfollowMakesUserUnfollowingAndRedirects()
-    {
-        $user = $this->login();
-        $owner_id = $this->create('user');
-        $followed_collection_dao = new models\dao\FollowedCollection();
-        $collection_id = $this->create('collection', [
-            'user_id' => $owner_id,
-            'type' => 'collection',
-            'is_public' => 1,
-        ]);
-        $this->create('followed_collection', [
-            'user_id' => $user->id,
-            'collection_id' => $collection_id,
-        ]);
-
-        $response = $this->appRun('post', "/collections/{$collection_id}/unfollow", [
-            'csrf' => $user->csrf,
-        ]);
-
-        $this->assertResponse($response, 302, "/collections/{$collection_id}");
-        $this->assertSame(0, $followed_collection_dao->count());
-    }
-
-    public function testUnfollowRedirectsIfNotConnected()
-    {
-        $user_id = $this->create('user');
-        $owner_id = $this->create('user');
-        $followed_collection_dao = new models\dao\FollowedCollection();
-        $collection_id = $this->create('collection', [
-            'user_id' => $owner_id,
-            'type' => 'collection',
-            'is_public' => 1,
-        ]);
-        $this->create('followed_collection', [
-            'user_id' => $user_id,
-            'collection_id' => $collection_id,
-        ]);
-
-        $response = $this->appRun('post', "/collections/{$collection_id}/unfollow", [
-            'csrf' => 'a token',
-        ]);
-
-        $this->assertResponse($response, 302, "/login?redirect_to=%2Fcollections%2F{$collection_id}");
-        $this->assertSame(1, $followed_collection_dao->count());
-    }
-
-    public function testUnfollowFailsIfCollectionDoesNotExist()
-    {
-        $user = $this->login();
-        $owner_id = $this->create('user');
-        $followed_collection_dao = new models\dao\FollowedCollection();
-        $collection_id = $this->create('collection', [
-            'user_id' => $owner_id,
-            'type' => 'collection',
-            'is_public' => 1,
-        ]);
-        $this->create('followed_collection', [
-            'user_id' => $user->id,
-            'collection_id' => $collection_id,
-        ]);
-
-        $response = $this->appRun('post', '/collections/unknown/unfollow', [
-            'csrf' => $user->csrf,
-        ]);
-
-        $this->assertResponse($response, 404);
-        $this->assertSame(1, $followed_collection_dao->count());
-    }
-
-    public function testUnfollowFailsIfUserHasNoAccess()
-    {
-        $user = $this->login();
-        $owner_id = $this->create('user');
-        $followed_collection_dao = new models\dao\FollowedCollection();
-        $collection_id = $this->create('collection', [
-            'user_id' => $owner_id,
-            'type' => 'collection',
-            'is_public' => 0,
-        ]);
-        $this->create('followed_collection', [
-            'user_id' => $user->id,
-            'collection_id' => $collection_id,
-        ]);
-
-        $response = $this->appRun('post', "/collections/{$collection_id}/unfollow", [
-            'csrf' => $user->csrf,
-        ]);
-
-        $this->assertResponse($response, 404);
-        $this->assertSame(1, $followed_collection_dao->count());
-    }
-
-    public function testUnfollowFailsIfCsrfIsInvalid()
-    {
-        $user = $this->login();
-        $owner_id = $this->create('user');
-        $followed_collection_dao = new models\dao\FollowedCollection();
-        $collection_id = $this->create('collection', [
-            'user_id' => $owner_id,
-            'type' => 'collection',
-            'is_public' => 1,
-        ]);
-        $this->create('followed_collection', [
-            'user_id' => $user->id,
-            'collection_id' => $collection_id,
-        ]);
-
-        $response = $this->appRun('post', "/collections/{$collection_id}/unfollow", [
-            'csrf' => 'not the token',
-        ]);
-
-        $this->assertResponse($response, 302, "/collections/{$collection_id}");
-        $this->assertFlash('error', 'A security verification failed: you should retry to submit the form.');
-        $this->assertSame(1, $followed_collection_dao->count());
-    }
 }
